PICKING WORKS BOYZ

+++ Source/Ursine3D/include/Core/Graphics/Core/DirectX/ShaderBuffers/ShaderBufferList.h
@@ -132,7 +132,7 @@ namespace ursine
 
         struct ComputeIDOutput
         {
-            float id;
+            unsigned id;
         };
     }
 }
